создан контроллер для отрисовки доски с заданиями
